+++ apm-agent-plugins/apm-neo4j-plugin/pom.xml
@@ -11,7 +11,7 @@
         <apm-agent-parent.base.dir>${project.basedir}/../..</apm-agent-parent.base.dir>
     </properties>
 
-    <artifactId>apm-neo4j3-plugin</artifactId>
+    <artifactId>apm-neo4j-plugin</artifactId>
     <name>${project.groupId}:${project.artifactId}</name>
 
     <dependencies>
