+++ apm-agent-plugins/apm-neo4j-plugin/src/main/resources/META-INF/services/co.elastic.apm.agent.bci.ElasticApmInstrumentation
@@ -1,0 +1,1 @@
+co.elastic.apm.agent.neo4j.StatementRunnerInstrumentation
