--- apm-agent-plugins/apm-neo4j3-plugin/src/main/resources/META-INF/services/co.elastic.apm.agent.bci.ElasticApmInstrumentation
@@ -1,2 +1,0 @@
-co.elastic.apm.agent.neo4j3.TransactionInstrumentation$CreateSpanInstrumentation
-co.elastic.apm.agent.neo4j3.TransactionInstrumentation$FinishSpanInstrumentation
