+++ apm-agent-plugins/pom.xml
@@ -44,7 +44,7 @@
         <module>apm-jmx-plugin</module>
         <module>apm-mule4-plugin</module>
         <module>apm-mongoclient-plugin</module>
-        <module>apm-neo4j3-plugin</module>
+        <module>apm-neo4j-plugin</module>
     </modules>
 
     <dependencies>
