+++ elastic-apm-agent/pom.xml
@@ -176,7 +176,7 @@
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>apm-neo4j3-plugin</artifactId>
+            <artifactId>apm-neo4j-plugin</artifactId>
             <version>${project.version}</version>
         </dependency>
